refactor stub utils from moe kernels
